pyproject.toml
[tool.poetry]
name = "aramis"
version = "0.1.0"
description = "Another NLP engine"
authors = ["Rémy Sanchez <remy.sanchez@hyperthese.net>"]
license = "AGPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.6"
cyhunspell = "^1.3.4"
scipy = "^1.4.1"
numpy = "^1.18.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pytest = "^5.3.3"
ipython = "^7.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
